Module Fixpoint.Reduce

module type GRAPHI = sig ... end

Interface for directed graphs with labeled vertices. The reduction * algorithm reduces graphs into a regex over the alphabet of edge values (be * it edge labels, or pairs of vertices if the edges are unlabeled). The input * graph must be without self-loops, i.e. edges of the form (n,n). All the nodes should also be reachable from the initial node.

module Make (G : GRAPHI) (R : Regex.S with type letter = G.E.t) : sig ... end