Module type Regex.S

Generic signature for a regex interface.

type letter

type t

val empty : t

val epsilon : t

val join : t -> t -> t

val append : t -> letter -> t

val append_star : t -> t -> t

val equal : t -> t -> bool

val hash : t -> int

val pretty : Format.formatter -> t -> unit

val pretty_reverse : Format.formatter -> t -> unit

Print a regex in reverse. Useful sometimes when a regex is too big to be * fully printed, and you only want to see the end, you can start printing it * in reverse and interrupt the program.