Module Wto.Make

This functor provides the partitioning algorithm constructing a WTO.

Parameters

module Node : sig ... end

Signature

type pref = Node.t -> Node.t -> int

Partial order of preference for the choice of the head of a loop. pref current_head new_candidate must return < 0 if new_candidate is preferred to current_head. Use "(fun _ _ -> 0)" for no specific preference.

val partition : pref:pref -> init:Node.t -> succs:(Node.t -> Node.t list) -> Node.t partition

Implements Bourdoncle "Efficient chaotic iteration strategies with widenings" algorithm to compute a WTO.

val pretty_partition : Format.formatter -> Node.t partition -> unit

val pretty_component : Format.formatter -> Node.t component -> unit

val equal_component : Node.t component -> Node.t component -> bool

val equal_partition : Node.t partition -> Node.t partition -> bool